Script victory condition
There is a bug where enabling the standard victory condition for Red on
Day 1 results in the allied team winning because they own all the towns.

As a workaround, we now enable the standard victory condition for Red
when they capture Dartmoor (the green town). This is fine because Red
can't win by conquest without capturing all towns, so at some point they
have to capture Dartmoor.