CoA: Give Teal's necromancers Advanced Necromancy
and fix bug where the hero army in Orange's starting area would move on
turn 1 when Orange was a human player - I think this is caused by a bug
in the game. The fix is to start the army unowned and change the owner
to Orange if Orange is a computer.